refactor: rename internal files to better reflect their specific responsibilities
